% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance_composite}
\alias{abundance_composite}
\title{Abundance composite score between CLEAN and DIRTY}
\usage{
abundance_composite(
  p_counts,
  d_counts,
  weights = c(spearman = 0.6, hellinger = 0.4)
)
}
\description{
Abundance composite score between CLEAN and DIRTY
}
\keyword{internal}
