% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_tables}
\alias{collapse_tables}
\title{Collapse DIRTY ASVs into CLEAN parents}
\usage{
collapse_tables(
  clean_tab,
  dirty_tab,
  mapping,
  clean_seqs,
  dirty_seqs,
  keep_unassigned = TRUE
)
}
\description{
Collapse DIRTY ASVs into CLEAN parents
}
\keyword{internal}
