% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence.R
\name{presence_composite}
\alias{presence_composite}
\title{Presence composite score between CLEAN and DIRTY}
\usage{
presence_composite(
  p_counts,
  d_counts,
  presence_cutoff = 1,
  weights = c(phi = 0.5, ppmi = 0.3, f1 = 0.2),
  min_recall = 0.5,
  min_n = 6
)
}
\description{
Presence composite score between CLEAN and DIRTY
}
\keyword{internal}
