% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{select_parents}
\alias{select_parents}
\title{Select CLEAN parents for each DIRTY ASV (advanced neighborhood + z-margin logic)}
\usage{
select_parents(
  identity_mat,
  mismatches_mat,
  single_score_mat,
  presence_score_mat,
  abundance_score_mat,
  recall_mat,
  phi01_mat,
  ppmi_tanh_mat,
  f1_mat,
  spearman01_mat,
  hellinger_mat,
  dominance_mean_ratio_mat,
  dominance_frac_exceed_mat,
  dominance_n_both_mat,
  dominance_pass_mat,
  min_identity,
  min_presence_score,
  enforce_dominance,
  identity_neighborhood_mismatches,
  neighborhood_z_margin,
  clean_tab
)
}
\description{
Select CLEAN parents for each DIRTY ASV (advanced neighborhood + z-margin logic)
}
\keyword{internal}
